Publish

+++ aspnet-core/src/Nguyen_Tan_Phat_Project.Application/Module/StructureAppService/EmployeeManagement/EmployeeAppService.cs
@@ -13,6 +13,7 @@ using Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement.dtos
 using Abp.Linq.Extensions;
 using Microsoft.EntityFrameworkCore;
 using Nguyen_Tan_Phat_Project.Authorization.Users;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
 {
@@ -21,12 +22,16 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
     {
         private IRepository<Employee, string> _employeeRepository;
         private IRepository<Structure, string> _structureRepository;
+        private readonly IRepository<Expenses, string> _expensesRepository;
         private IRepository<User, long> _userReposistory;
+        private readonly IRepository<ExportImport, string> _exportImportRepository;
         private IRepository<BankAccount> _bankRepository;
         private IRepository<CMND> _cmndRepository;
     
         public EmployeeAppService(IRepository<Employee, string> employeeRepository
             , IRepository<Structure, string> structureRepository
+            , IRepository<Expenses, string> expensesRepository
+            , IRepository<ExportImport, string> exportImportRepository
             , IRepository<BankAccount> bankRepository
             , IRepository<User, long> userRepository
             , IRepository<CMND> cmndRepository)
@@ -36,6 +41,8 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
             _bankRepository = bankRepository;
             _userReposistory = userRepository;
             _cmndRepository = cmndRepository;
+            _expensesRepository = expensesRepository;
+            _exportImportRepository = exportImportRepository;
         }
 
         [AbpAuthorize(PermissionNames.Page_System_Employee_Add)]
@@ -62,6 +69,7 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
                         TaxIdentification = input.TaxIdentification,
                         phoneNumber = input.PhoneNumber,
                         EmployeeSalary = input.EmployeeSalary,
+                        EmployeeAllowance = input.EmployeeAllowance,
                         SalaryFactor = input.SalaryFactor,
                         TypeOfContract = input.TypeOfContract,
                     };
@@ -78,6 +86,7 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
                         TaxIdentification = input.TaxIdentification,
                         phoneNumber = input.PhoneNumber,
                         EmployeeSalary = input.EmployeeSalary,
+                        EmployeeAllowance = input.EmployeeAllowance,
                         SalaryFactor = input.SalaryFactor,
                         TypeOfContract = input.TypeOfContract,
                         BankAccount = input.employeeBankAccount
@@ -108,9 +117,11 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
             try
             {
                 var employeeAccount = await _userReposistory.FirstOrDefaultAsync(e => e.UserName == id);
-                if (employeeAccount != null)
+                var employeeInDon = await _exportImportRepository.FirstOrDefaultAsync(e => e.OrderCreator == id);
+                var employeeInDon2 = await _expensesRepository.FirstOrDefaultAsync(e => e.OrderCreator == id);
+                if (employeeAccount != null || employeeInDon != null || employeeInDon2 != null)
                 {
-                    throw new UserFriendlyException("Nhân viên có tài khoản không thể bị xóa");
+                    throw new UserFriendlyException("Nhân viên không thể bị xóa");
                 }
                 var employee = await _employeeRepository.GetAsync(id);
                 await _cmndRepository.DeleteAsync(e => e.Employee.Id == id);
@@ -165,6 +176,7 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
                 employeeDto.WorkUnit = _structureRepository.Get(input.WorkUnit);
                 employeeDto.TaxIdentification = input.TaxIdentification;
                 employeeDto.EmployeeSalary = input.EmployeeSalary;
+                employeeDto.EmployeeAllowance = input.EmployeeAllowance;
                 employeeDto.SalaryFactor = input.SalaryFactor;
                 employeeDto.phoneNumber = input.PhoneNumber;
                 employeeDto.TypeOfContract = input.TypeOfContract;
@@ -175,6 +187,13 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
                     await _bankRepository.InsertAsync(input.employeeBankAccount);
                     employeeDto.BankId = input.employeeBankAccount.BankId;
                     await _employeeRepository.UpdateAsync(employeeDto);
+                } else
+                {
+                    var bankAccount = await _bankRepository.FirstOrDefaultAsync(e => e.BankId == input.employeeBankAccount.BankId);
+                    bankAccount.BankName = input.employeeBankAccount.BankName;
+                    bankAccount.BankCity = input.employeeBankAccount.BankCity;
+                    bankAccount.BankAddress = input.employeeBankAccount.BankAddress;
+                    await _bankRepository.UpdateAsync(bankAccount);
                 }
 
                 //_cmndRepository.Update(input.EmployeeCMND);
@@ -199,11 +218,18 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
                     JobTitle = e.JobTitle,
                     WorkUnit = e.WorkUnit.UnitName,
                     EmployeePhone = e.phoneNumber,
+                    EmployeeAllowance = e.EmployeeAllowance,
+                    EmployeeSalary = e.EmployeeSalary,
                     AccountId = e.BankAccount.BankId,
                     AccountName = e.BankAccount.BankName,
                 }).PageBy(input).ToListAsync();
 
                 int totalCount = _employeeRepository.Count();
+                
+                if (!string.IsNullOrEmpty(input.UnitCode))
+                {
+                    totalCount = _employeeRepository.GetAll().Where(e => e.WorkUnitId.Contains(input.UnitCode)).Count();
+                }
 
                 return new PagedResultDto<EmployeeGetAllDto>
                 {
@@ -246,6 +272,22 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
 
             return result;
         }
+        
+        public async Task<EmployeeSelectForAccountList> GetEmployeeSelectWithStructureId(string structureId)
+        {
+            var query = await _employeeRepository.GetAll()
+                .Where(e => e.WorkUnitId == structureId)
+                .Select(e => new EmployeeSelectForAccount
+                {
+                    Code = e.Id,
+                    Name = e.EmployeeName,
+                }).ToListAsync();
+
+            EmployeeSelectForAccountList result = new EmployeeSelectForAccountList();
+            result.items = query;
+
+            return result;
+        }
 
         public async Task<EmployeeOutputDto> GetAsync(string id)
         {
@@ -261,6 +303,7 @@ namespace Nguyen_Tan_Phat_Project.Module.StructureAppService.EmployeeManagement
                 TaxIdentification = query.TaxIdentification,
                 EmployeePhone = query.phoneNumber,
                 EmployeeSalary = query.EmployeeSalary,
+                EmployeeAllowance = query.EmployeeAllowance,
                 SalaryFactor = query.SalaryFactor,
                 TypeOfContract = query.TypeOfContract,
                 EmployeeBankAccount = _bankRepository.GetAll().FirstOrDefault(b => b.BankId == query.BankId),
